+++ js/util.js
@@ -1,8 +1,6 @@
-import './util.js';
-/*import './data.js';*/
+import {AVATARS, TYPES, CHECKINS, CHECKOUTS, FEATURES, PHOTOS, SIMILAR_HOTELS_COUNT} from './data.js'
 
-
-/*function getRandInt(minNumber, maxNumber) {
+function getRandInt(minNumber, maxNumber) {
   if(minNumber > maxNumber || maxNumber === minNumber || minNumber < 0 || maxNumber < 0) {
     throw new Error('Error description');
   }
@@ -15,56 +13,6 @@ function getRandFloat(minNumber, maxNumber, float) {
   }
   return Number((Math.random() * (maxNumber - minNumber + 1) + minNumber).toFixed(float));
 }
-
-const AVATARS = [
-  'img/avatars/user01.png',
-  'img/avatars/user02.png',
-  'img/avatars/user03.png',
-  'img/avatars/user04.png',
-  'img/avatars/user05.png',
-  'img/avatars/user06.png',
-  'img/avatars/user07.png',
-  'img/avatars/user08.png',
-  'img/avatars/user09.png',
-  'img/avatars/user10.png',
-];
-
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-
-const CHECKINS = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const CHECKOUTS = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-
-const SIMILAR_HOTELS_COUNT = 10;
 
 const getRandomArrayElement = (elements) => elements[getRandInt(0, elements.length - 1)];
 
@@ -99,4 +47,4 @@ const generateHotel = () => ({
 
 const similarHotels = new Array(SIMILAR_HOTELS_COUNT).fill(null).map(() => generateHotel());
 
-console.log(similarHotels);*/
+console.log(similarHotels);
